pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rln-rag"
version = "0.1.0"
description = "LangGraph based RAG playground for Llama 3.1 8B"
requires-python = ">=3.11"
authors = [{ name = "sho" }]
readme = "README.md"
dependencies = [
    "langchain==1.0.7",
    "langgraph==1.0.3",
    "langchain-community==0.4.1",
    "langchain-text-splitters==1.0.0",
    "langchain-openai==1.0.3",
    "langchain-groq==1.0.1",
    "faiss-cpu==1.12.0",
    "python-dotenv==1.2.1",
    "pydantic==2.12.4",
    "typer==0.20.0",
    "rich==14.2.0",
    "sentence-transformers==5.1.2",
    "ollama==0.6.1",
    "openai==2.8.0",
    "transformers==4.57.1",
    "torch==2.9.1",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "tokenizers==0.22.1",
    "safetensors==0.6.2",
]

[project.optional-dependencies]
dev = [
    # langgraph dev(inmem) 서버용
    "langgraph-cli[inmem]>=0.4.7",
]

[project.scripts]
rln-rag = "rln_rag.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
